Permission Checked Added in Banner Controller

+++ src/controllers/banner/deleteBannerImageController.js
@@ -1,11 +1,20 @@
 // ALL REQUIRES
 const { deleteBannerImage } = require("../../helpers/bannerHelper");
 const { deleteBannerImageRequest } = require("../../requests/bannerRequest");
+const { checkPermission } = require("../../utils/permissionChecker");
 const { singleResponse } = require("../../utils/response");
 
 
 // CONTROLLER
 module.exports = async (req, db, user, isAuth, TENANTID) => {
+    // Permission Name of this API
+    const permissionName = "banner";
+    // Check Permission
+    const checkPermissions = await checkPermission(db, user, TENANTID, permissionName);
+    if (!checkPermissions.success) {
+        return { message: "You dont have access to this route, please contact support to have you give this route permission!!!", status: false };
+    }
+
     // Validate DELETE BANNER IMAGE Request
     const validate = await deleteBannerImageRequest(req);
     if (!validate.success) {
